[pc-61] Split session controller acceptance into several files
Only moving code, no edition
